+++ packages/core/src/application/application_ref.ts
@@ -12,7 +12,6 @@ import {setThrowInvalidWriteToSignalError} from '@angular/core/primitives/signal
 import {Observable} from 'rxjs';
 import {first, map} from 'rxjs/operators';
 
-import {getCompilerFacade, JitCompilerUsage} from '../compiler/compiler_facade';
 import {Console} from '../console';
 import {inject} from '../di';
 import {Injectable} from '../di/injectable';
@@ -20,23 +19,18 @@ import {InjectionToken} from '../di/injection_token';
 import {Injector} from '../di/injector';
 import {EnvironmentInjector} from '../di/r3_injector';
 import {ErrorHandler, INTERNAL_APPLICATION_ERROR_HANDLER} from '../error_handler';
-import {formatRuntimeError, RuntimeError, RuntimeErrorCode} from '../errors';
+import {RuntimeError, RuntimeErrorCode, formatRuntimeError} from '../errors';
 import {Type} from '../interface/type';
-import {COMPILER_OPTIONS, CompilerOptions} from '../linker/compiler';
 import {ComponentFactory, ComponentRef} from '../linker/component_factory';
 import {ComponentFactoryResolver} from '../linker/component_factory_resolver';
-import {NgModuleFactory, NgModuleRef} from '../linker/ng_module_factory';
+import {NgModuleRef} from '../linker/ng_module_factory';
 import {ViewRef} from '../linker/view_ref';
-import {isComponentResourceResolutionQueueEmpty, resolveComponentResources} from '../metadata/resource_loading';
 import {PendingTasks} from '../pending_tasks';
 import {AfterRenderEventManager} from '../render3/after_render_hooks';
-import {assertNgModuleType} from '../render3/assert';
 import {ComponentFactory as R3ComponentFactory} from '../render3/component_ref';
 import {isStandalone} from '../render3/definition';
-import {ChangeDetectionMode, detectChangesInternal, MAXIMUM_REFRESH_RERUNS} from '../render3/instructions/change_detection';
+import {ChangeDetectionMode, MAXIMUM_REFRESH_RERUNS, detectChangesInternal} from '../render3/instructions/change_detection';
 import {FLAGS, LView, LViewFlags} from '../render3/interfaces/view';
-import {setJitOptions} from '../render3/jit/jit_options';
-import {NgModuleFactory as R3NgModuleFactory} from '../render3/ng_module_ref';
 import {publishDefaultGlobalUtils as _publishDefaultGlobalUtils} from '../render3/util/global_utils';
 import {requiresRefreshOrTraversal} from '../render3/util/view_utils';
 import {ViewRef as InternalViewRef} from '../render3/view_ref';
@@ -59,55 +53,6 @@ import {ApplicationInitStatus} from './application_init';
 export const APP_BOOTSTRAP_LISTENER =
     new InjectionToken<ReadonlyArray<(compRef: ComponentRef<any>) => void>>(
         ngDevMode ? 'appBootstrapListener' : '');
-
-export function compileNgModuleFactory<M>(
-    injector: Injector, options: CompilerOptions,
-    moduleType: Type<M>): Promise<NgModuleFactory<M>> {
-  ngDevMode && assertNgModuleType(moduleType);
-
-  const moduleFactory = new R3NgModuleFactory(moduleType);
-
-  // All of the logic below is irrelevant for AOT-compiled code.
-  if (typeof ngJitMode !== 'undefined' && !ngJitMode) {
-    return Promise.resolve(moduleFactory);
-  }
-
-  const compilerOptions = injector.get(COMPILER_OPTIONS, []).concat(options);
-
-  // Configure the compiler to use the provided options. This call may fail when multiple modules
-  // are bootstrapped with incompatible options, as a component can only be compiled according to
-  // a single set of options.
-  setJitOptions({
-    defaultEncapsulation: _lastDefined(compilerOptions.map(opts => opts.defaultEncapsulation)),
-    preserveWhitespaces: _lastDefined(compilerOptions.map(opts => opts.preserveWhitespaces)),
-  });
-
-  if (isComponentResourceResolutionQueueEmpty()) {
-    return Promise.resolve(moduleFactory);
-  }
-
-  const compilerProviders = compilerOptions.flatMap((option) => option.providers ?? []);
-
-  // In case there are no compiler providers, we just return the module factory as
-  // there won't be any resource loader. This can happen with Ivy, because AOT compiled
-  // modules can be still passed through "bootstrapModule". In that case we shouldn't
-  // unnecessarily require the JIT compiler.
-  if (compilerProviders.length === 0) {
-    return Promise.resolve(moduleFactory);
-  }
-
-  const compiler = getCompilerFacade({
-    usage: JitCompilerUsage.Decorator,
-    kind: 'NgModule',
-    type: moduleType,
-  });
-  const compilerInjector = Injector.create({providers: compilerProviders});
-  const resourceLoader = compilerInjector.get(compiler.ResourceLoader);
-  // The resource loader can also return a string while the "resolveComponentResources"
-  // always expects a promise. Therefore we need to wrap the returned value in a promise.
-  return resolveComponentResources(url => Promise.resolve(resourceLoader.get(url)))
-      .then(() => moduleFactory);
-}
 
 export function publishDefaultGlobalUtils() {
   ngDevMode && _publishDefaultGlobalUtils();
@@ -732,16 +677,6 @@ export function remove<T>(list: T[], el: T): void {
     list.splice(index, 1);
   }
 }
-
-function _lastDefined<T>(args: T[]): T|undefined {
-  for (let i = args.length - 1; i >= 0; i--) {
-    if (args[i] !== undefined) {
-      return args[i];
-    }
-  }
-  return undefined;
-}
-
 
 let whenStableStore: WeakMap<ApplicationRef, Promise<void>>|undefined;
 /**
